- added the label builder to the tests - created a general entry point to run all tests
