+++ test/test_main.dart
@@ -1,0 +1,7 @@
+import 'widgets/dynamic_timeline_tests.dart';
+import 'widgets/timeline_item_tests.dart';
+
+void main(){
+  DynamicTimelineTests.run();
+  TimelineItemTests.run();
+}
